Restructured files and added pydantic base model for raw log
